Centralize stream definition
