+++ libs/@local/hash-backend-utils/src/realtime.ts
@@ -1,3 +1,9 @@
+import { getRequiredEnv } from "./environment";
+import { Logger } from "./logger";
+import { Entity, EntityType, PropertyType } from "./pg-tables";
+import { RedisConfig, setupRedisClient } from "./redis";
+import { RedisStreamConsumer, RedisStreamProducer } from "./stream/redis";
+
 /**
  * @todo Consider adding realtime handling for types
  *   https://app.asana.com/0/0/1202922776289399/f
@@ -41,3 +47,63 @@ export const isSupportedRealtimePropertyTypeTable = (
   supportedTables.propertyTypeTables.includes(
     table as SupportedRealtimePropertyTypeTable,
   );
+
+const getStreams = () => {
+  return {
+    entityStream: getRequiredEnv("HASH_REALTIME_ENTITY_STREAM_NAME"),
+    entityTypeStream: getRequiredEnv("HASH_REALTIME_ENTITY_TYPE_STREAM_NAME"),
+    propertyTypeStream: getRequiredEnv(
+      "HASH_REALTIME_PROPERTY_TYPE_STREAM_NAME",
+    ),
+  };
+};
+
+// The realtime service will push all updates from the Postgres changestream to the
+// following queues.
+export const generateStreamProducers = (
+  logger: Logger,
+  redisConfig: RedisConfig,
+) => {
+  const streams = getStreams();
+  return {
+    entityStream: new RedisStreamProducer<Entity>(
+      logger,
+      setupRedisClient(logger, redisConfig),
+      streams.entityStream,
+    ),
+    entityTypeStream: new RedisStreamProducer<EntityType>(
+      logger,
+      setupRedisClient(logger, redisConfig),
+      streams.entityTypeStream,
+    ),
+    propertyTypeStream: new RedisStreamProducer<PropertyType>(
+      logger,
+      setupRedisClient(logger, redisConfig),
+      streams.propertyTypeStream,
+    ),
+  } as const;
+};
+
+export const generateStreamConsumers = (
+  logger: Logger,
+  redisConfig: RedisConfig,
+) => {
+  const streams = getStreams();
+  return {
+    entityStream: new RedisStreamConsumer<Entity>(
+      logger,
+      setupRedisClient(logger, redisConfig),
+      streams.entityStream,
+    ),
+    entityTypeStream: new RedisStreamConsumer<EntityType>(
+      logger,
+      setupRedisClient(logger, redisConfig),
+      streams.entityTypeStream,
+    ),
+    propertyTypeStream: new RedisStreamProducer<PropertyType>(
+      logger,
+      setupRedisClient(logger, redisConfig),
+      streams.propertyTypeStream,
+    ),
+  } as const;
+};
